refactor: type 추가와 파일이동
